requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.4.1
async-timeout==4.0.1
attrs==21.2.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.8
click==8.0.3
Django==3.2.9
Flask==2.0.2
frozenlist==1.2.0
google-api-core==2.2.2
google-auth==2.3.3
google-cloud-dialogflow==2.10.0
googleapis-common-protos==1.53.0
grpcio==1.42.0
grpcio-status==1.42.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
multidict==5.2.0
proto-plus==1.19.8
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyee==8.2.2
pytz==2021.3
requests==2.26.0
rsa==4.8
six==1.16.0
slackclient==2.9.3
slackeventsapi==3.0.0
sqlparse==0.4.2
typing_extensions==4.0.0
urllib3==1.26.7
Werkzeug==2.0.2
yarl==1.7.2
